requirements.txt
fastapi
uvicorn
qdrant-client==1.9.0
trafilatura
openai>=1.0.0
pydantic
python-dotenv
requests
beautifulsoup4